refactor: Consolidate `getAll` and `getEntry` methods into simply `get` handling both cases
Based on the parameters passed into the method, we can figure out whether it is a `getAll` or `getEntry` method.
This ensures a smaller public API, which should be easier to learn. We should handle the complexity rather within
the methods and expose a simple API.

Note: There are some code duplication now, but leaving the duplication for now since it will be easier to maintain,
we can refactor later if we feel like it.